requirements.txt
annotated-types==0.5.0
anyio==3.7.1
cached-property==1.5.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.2
Flask==2.2.5
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.7
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.4
MarkupSafe==2.1.5
openai==1.35.14
Pillow==9.5.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==0.21.1
requests==2.31.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
zipp==3.15.0
